added new attributes to the csv file and adjusted the CardCSVReader class
To handle the new CSV format for reading cards, the CardCSVReader class was updated to accommodate the different types of cards: PayCard, MoveCard, and JokerCard. Each card type is identified by the CardType enum, and specific attributes are parsed accordingly. For MoveCard, additional attributes such as the field ID were introduced to correctly represent the functionality of the card, since you only get money if you have gone past the start.