added methods for bank and risiko cards (pay, move) and implemented helper methods to calculate the detination of move
In the GameBoardViewModel I had to add a function in the handleAction() so that the Update_Money action can be carried out correctly, because previously you could no longer continue playing correctly after you got to the start. I also added two methods payForCard() and moveForCard(). The pay method simply updates the money of the current player for all players. The move method calculates the number of fields the player has to move to get to the desired field.
In the GameBoard I had to change a small thing in the checkEndFieldPosition() to ensure that you only get the money from the start once per round. I also adjusted showCardRisiko/Bank() so that they now call the correct methods (pay or move).
In the game I created two methods to get the number of fields and their position. And you also had to set the player to isOnTurn() = true.