implemented RECONNECT_DISCARD request
A Player is now be able to discard a reconnection to a previous joined game. The player is removed from the game list and all fields realted to this player are unmarked in the UI and the Owner is set to null so all the fields get ownable for all other players. The Character Figure on the UI gets also removed and the Figures of the other Players will switch Position in right order relatively to the Field Imageview.