added method to set player to default
When a Player leaves a playing game all Stats/Variables gets set to defautl values so on a reconnect the game starts with expected values for the position, money and so on. That is the case for the leaving Player itself if he leaves the game or from the host that kicks a Player.