added methods to add joker cards
I have extended showCardRisiko() in the GameBoard so that it now checks whether the drawn risiko card is a joker. If so, a joker card is assigned to the corresponding player. In the GameBoardViewModel I then created two methods to add the jokers. One for the current player and one for the others. I also adapted the GameBoardAction accordingly and the joker class.