added support for joker cards
I added a new array list of joker card to the Player and furthermore Getter, Setter and methods to add, delete cards and to check if the player has joker cards. In the PLayerItemAdapter a new TextView was added to the ViewHolder to visualize how many joker cards a player has. Therefore, I had to adapt the player_stat_item.xml.